+++ semcore/typography/src/index.d.ts
@@ -2,8 +2,7 @@ import React from 'react';
 import { PropGetterFn, UnknownProperties, Intergalactic } from '@semcore/core';
 
 import { Property } from 'csstype';
-import { Box, BoxProps, Flex, FlexProps } from '@semcore/flex-box';
-import { KeyboardFocusProps } from '@semcore/core/lib/utils/enhances/keyboardFocusEnhance';
+import { BoxProps, Flex, FlexProps } from '@semcore/flex-box';
 
 /** @deprecated */
 export interface ITextProps extends TextProps, UnknownProperties {}